backend/requirements.txt
asgiref==3.7.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==21.3
payu-python==0.1.5
pexpect==4.8.0
Pillow==10.2.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.27
python3-openid==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==2.0.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.4.4
ssh-import-id==5.11
stripe==12.2.0
systemd-python==234
Twisted==22.1.0
typing_extensions==4.9.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
django-storages==1.14.2
boto3==1.34.34
